requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.1
ffmpeg==1.4
Flask==1.1.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pydub==0.23.1
pylint==2.4.4
pytube3==9.6.4
six==1.14.0
SpeechRecognition==3.8.1
SQLAlchemy==1.3.16
typed-ast==1.4.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1
